server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "websockets>=16.0",
    "aiokafka>=0.12.0",
    "redis>=5.2.0",
    "python-socketio>=5.16.1",
    "motor>=3.7.1",
]
